RangeLock: Add capability to lock multiple ranges (#54)
Ability to lock multiple ranges is required in BulkIO APIs

To avoid deadlock, the range of block offsets are sorted
It is the responsibility of caller to ensure coinciding ranges
are not passed. It will cause deadlock.
Additionally, debug build, contains code to detect coinciding
range is added.

If there are 2 ranges to lock, we locked the first range, and
if we couldn't lock the second range - we do not release the
lock for first range.

LockGuard API ensures
---
Given list of ranges - locks are always taken in increasing order.
Therefore, in the constructor of LockGuard(), there is check for
std::is_sorted(), if it is not already sorted we sort the range.
Deadlock is avoided as, multiple racing calls to Lock, always
take lock in same order. For us the order is increasing order.
As long as, ranges are locked only in increasing order, every
lock should progress.

LockGuard API assumes
---
the caller will not pass coinciding ranges to lock. If coinciding
ranges are passed, it will surely deadlock. Again, in the
constructor, with DEBUG compilation, you will see check for
coinciding ranges. We throw and exception if range coincides.

There is a TEST specifically, added to to verify LockGuard
API deadlocks or throw exception if coinciding ranges are passed
for locking.

Signed-off-by: Prasad Joshi Prasad.Joshi@primaryio.com