Some fixes for Write Pipe Listener error (DP-1734 ) (#86)
Temporarily avoid using aeroSpike client internal retry for transient failures.
Retries in case of most of the failures are anyway explicitly done by stord.
We are hitting an issue with aerospike write pipeline API (which
may be related to a BUG in AeroSpike client code) where Pipe Listener callback is
being called multiple times in context of a single command. As suggestion at the
aerospike forum is to avoid using the client internal retries and with this change
we are not seeing Pipe Listener being called multiple times. We are looking
at the aerospike code currently to find more on this.

Link:- https://discuss.aerospike.com/t/c-client-crash-after-4-2-0/4827