Implement the ARM info rest call
The arm sync rest call to give the status
of each vmdk in the process of sync. Does
not give all the data as per spec but enough
that we can calculate the progress of the
sync.

Note:- This code has not been compiled
since I could not get the build to work correctly.

Signed-off-by: Anand Mitra <mitra@primaryio.com>